pyproject.toml
[project]
name = "cikgu"
version = "0.1.0"
description = "Online learning platform for Malaysia's SPM lessons"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.5",
    "cryptography>=46.0.1",
    "flask>=3.0.0",
    "python-dotenv>=1.0.0",
]
